🪲(fix) : fix minor deployment errors
