+++ app/page.tsx
@@ -22,18 +22,7 @@ import { CarouselCard } from "@/components/CarouselCards";
 import { ClientSection } from "@/components/ClientSection";
 import { StickyScroll } from "@/components/ui/sticky-scroll-reveal";
 import { PricingWithSwitch } from "@/components/ui/pricing-with-switch";
-import {
-  Camera,
-  Lightbulb,
-  Clapperboard,
-  BarChart,
-  Target,
-  Users,
-  Video,
-  ImageIcon,
-  Film,
-  LucideIcon,
-} from "lucide-react";
+import { services } from "@/lib/services";
 import BannerVideo from "@/components/ui/BannerVideo";
 import ServiceSection from "@/components/ServiceSection";
 const WorldMapSection = dynamic(
@@ -54,117 +43,6 @@ const tabColors = {
   socialmedia: "#992D47",
 };
 
-type Feature = {
-  name: string;
-  description: string;
-  icon: LucideIcon;
-};
-
-export type Service = {
-  eyebrow: string;
-  title: string;
-  description: string;
-  id: string;
-  image: string;
-  link: string;
-  reverse?: boolean;
-  features: Feature[];
-  color: string;
-};
-
-export const services: Service[] = [
-  {
-    eyebrow: "Studio de tournage",
-    title: "Des espaces pensés pour vos productions",
-    color: "text-indigo-400",
-    link: "studio-de-tournage",
-    id: "studio-de-tournage",
-    description:
-      "Découvrez nos espaces, équipements et options de location pour vos productions. Nous mettons à disposition un environnement professionnel conçu pour maximiser la qualité de vos contenus tout en simplifiant vos tournages.",
-    image: "/img/studio.jpg",
-    features: [
-      {
-        name: "Espace modulable.",
-        description:
-          "Un studio adaptable pour différents formats de tournage, du contenu social media aux productions plus ambitieuses.",
-        icon: Camera,
-      },
-      {
-        name: "Éclairage professionnel.",
-        description:
-          "Des solutions d’éclairage optimisées pour garantir un rendu visuel cohérent et esthétique.",
-        icon: Lightbulb,
-      },
-      {
-        name: "Matériel disponible.",
-        description:
-          "Accédez à l’équipement essentiel pour produire efficacement sans contraintes techniques.",
-        icon: Clapperboard,
-      },
-    ],
-  },
-  {
-    eyebrow: "Accompagnement stratégique",
-    title: "Une vision claire pour votre croissance",
-    color: "text-fuchsia-400",
-    link: "accompagnement-strategique",
-    id: "accompagnement-strategique",
-    description:
-      "Community management, audit digital et stratégie sur mesure pour votre croissance. Nous analysons votre présence actuelle et construisons des actions concrètes pour améliorer votre visibilité et vos performances.",
-    image: "/img/Stratégie.webp",
-    reverse: true,
-    features: [
-      {
-        name: "Audit digital.",
-        description:
-          "Analyse complète de votre présence en ligne pour identifier les opportunités d’amélioration.",
-        icon: BarChart,
-      },
-      {
-        name: "Stratégie sur mesure.",
-        description:
-          "Des recommandations adaptées à votre marché et à vos objectifs business.",
-        icon: Target,
-      },
-      {
-        name: "Community management.",
-        description:
-          "Gestion et animation de vos réseaux sociaux pour créer une audience engagée.",
-        icon: Users,
-      },
-    ],
-  },
-  {
-    eyebrow: "Audiovisuel",
-    title: "Donnez vie à votre image",
-    color: "text-cyan-400",
-    link: "audiovisuel",
-    id: "audiovisuel",
-    description:
-      "Prestations vidéo/photo, tournage, montage et livraison multi-formats. Nous vous accompagnons de la conception à la livraison pour produire du contenu impactant et adapté à tous vos canaux.",
-    image: "/img/Production.webp",
-    features: [
-      {
-        name: "Production vidéo.",
-        description:
-          "Tournage professionnel pour vos contenus marketing, corporate ou réseaux sociaux.",
-        icon: Video,
-      },
-      {
-        name: "Photographie.",
-        description:
-          "Des visuels de qualité pour valoriser votre marque et vos produits.",
-        icon: ImageIcon,
-      },
-      {
-        name: "Montage & formats.",
-        description:
-          "Optimisation et déclinaison de vos contenus pour toutes les plateformes.",
-        icon: Film,
-      },
-    ],
-  },
-];
 export default function Home() {
   const [bgColor, setBgColor] = useState("#000000");
   const [activeTab, setActiveTab] = useState("combo");
